Add threashold and move components
