+++ src/Graph/LineGraph/Line.js
@@ -2,7 +2,7 @@ import React, { useEffect, createRef, useMemo } from 'react';
 import * as d3 from 'd3';
 import PropTypes from 'prop-types';
 import * as C from './LineGraph.styled';
-import ClipPath from '../ClipPath';
+import ClipPath from '../components/ClipPath';
 
 const propTypes = {
   data: PropTypes.instanceOf(Array).isRequired,
@@ -18,7 +18,14 @@ const propTypes = {
 const defaultPtops = {};
 
 const Line = ({
-  xScale, yScale, data, yProp, xProp, dimensions, duration, updateTick,
+  xScale,
+  yScale,
+  data,
+  yProp,
+  xProp,
+  dimensions,
+  duration,
+  updateTick,
 }) => {
   const ref = createRef();
   const line = useMemo(() => d3.line()
@@ -30,14 +37,12 @@ const Line = ({
     // On first update-tick we dont want any duration/transition
     if (updateTick === 0) {
       d3.select(ref.current)
-        .select('.line')
         .datum(data)
         .attr('d', line);
     // On the upcomming ticks the user will request other
     // domains and we want to use duration/transition
     } else if (updateTick !== 0) {
       d3.select(ref.current)
-        .select('.line')
         .datum(data)
         .transition()
         .duration(duration)
@@ -47,8 +52,8 @@ const Line = ({
 
   return (
     <ClipPath dimensions={dimensions}>
-      <g ref={ref}>
-        <C.Line className="line" />
+      <g>
+        <C.Line ref={ref} className="line" />
       </g>
     </ClipPath>
   );
