+++ src/Graph/LineGraph/LineGraph.styled.js
@@ -5,3 +5,11 @@ export const Line = styled.path`
     stroke: steelblue;
     stroke-width: 1px;
 `;
+
+export const Threashold = styled.line`
+    fill: none;
+    stroke: ${({ theme }) => theme.ruby800};
+    stroke-dasharray: 5 3;
+    stroke-width: 1px;
+    
+`;
