+++ src/Graph/components/ClipPath/ClipPath.js
