+++ src/Graph/components/ClipPath/index.js
